requirements.txt
streamlit
opencv-contrib-python-headless
numpy
io
pillow